コンポーネント内で使われるカスタムフックをモックして呼び出してテストするようにしました。
